+++ tests/autoarm/test_example_configs.py
@@ -99,10 +99,9 @@ async def test_supplemental_config_installed(
     enquire_config = await hass.services.async_call(
         "autoarm", "enquire_configuration", None, blocking=True, return_response=True
     )
-    assert enquire_config["notify"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
-    assert enquire_config["notify"]["normal"]["source"] == ["calendar"]
-    assert enquire_config["notify"]["common"]["service"] == "notify.supernotify"
-    assert enquire_config["notify"]["common"]["supernotify"]
+    assert enquire_config["notify"]["service"] == "notify.send_message"
+    assert enquire_config["notify"]["profiles"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
+    assert enquire_config["notify"]["profiles"]["normal"]["source"] == ["calendar"]
     assert enquire_config["rate_limit"]["period"] == "60 seconds"
 
     await _reload_and_verify(hass, monkeypatch)
@@ -139,10 +138,10 @@ async def test_legacy_config_installed(
     enquire_config = await hass.services.async_call(
         "autoarm", "enquire_configuration", None, blocking=True, return_response=True
     )
-    assert enquire_config["notify"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
-    assert enquire_config["notify"]["normal"]["source"] == ["calendar"]
-    assert enquire_config["notify"]["common"]["service"] == "notify.supernotify"
-    assert enquire_config["notify"]["common"]["supernotify"]
+    assert enquire_config["notify"]["service"] == "notify.supernotify"
+    assert enquire_config["notify"]["profiles"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
+    assert enquire_config["notify"]["profiles"]["normal"]["source"] == ["calendar"]
+    assert enquire_config["notify"]["profiles"]["common"]["supernotify"]
     assert enquire_config["rate_limit"]["period"] == "60 seconds"
 
     await _reload_and_verify(hass, monkeypatch)
@@ -179,10 +178,10 @@ async def test_legacy_config_fresh_install(
     enquire_config = await hass.services.async_call(
         "autoarm", "enquire_configuration", None, blocking=True, return_response=True
     )
-    assert enquire_config["notify"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
-    assert enquire_config["notify"]["normal"]["source"] == ["calendar"]
-    assert enquire_config["notify"]["common"]["service"] == "notify.supernotify"
-    assert enquire_config["notify"]["common"]["supernotify"]
+    assert enquire_config["notify"]["service"] == "notify.supernotify"
+    assert enquire_config["notify"]["profiles"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
+    assert enquire_config["notify"]["profiles"]["normal"]["source"] == ["calendar"]
+    assert enquire_config["notify"]["profiles"]["common"]["supernotify"]
     assert enquire_config["rate_limit"]["period"] == "60 seconds"
 
     await _reload_and_verify(hass, monkeypatch)